Re-use account related functions in interchain gas payment indexing
