+++ rust/main/chains/hyperlane-sealevel/src/mailbox.rs
@@ -58,6 +58,7 @@ use solana_transaction_status::{
 };
 use tracing::{debug, info, instrument, warn};
 
+use crate::account::{search_accounts_by_discriminator, search_and_validate_account};
 use crate::error::HyperlaneSealevelError;
 use crate::transaction::search_dispatched_message_transactions;
 use crate::utils::{decode_h256, decode_h512, from_base58};
@@ -664,16 +665,22 @@ impl SealevelMailboxIndexer {
         nonce: u32,
     ) -> ChainResult<(Indexed<HyperlaneMessage>, LogMeta)> {
         let discriminator = hyperlane_sealevel_mailbox::accounts::DISPATCHED_MESSAGE_DISCRIMINATOR;
+        let nonce_bytes = nonce.to_le_bytes();
         let offset = 1 + 8 + 4 + 8; // the offset to get the `unique_message_pubkey` field
         let length = 32; // the length of the `unique_message_pubkey` field
-        let accounts = self
-            .search_accounts_by_discriminator(&discriminator, nonce, offset, length)
-            .await?;
+        let accounts = search_accounts_by_discriminator(
+            self.rpc(),
+            &self.program_id,
+            &discriminator,
+            &nonce_bytes,
+            offset,
+            length,
+        )
+        .await?;
 
-        let valid_message_storage_pda_pubkey = self
-            .search_and_validate_account(accounts, |account| {
-                self.dispatched_message_account(&account)
-            })?;
+        let valid_message_storage_pda_pubkey = search_and_validate_account(accounts, |account| {
+            self.dispatched_message_account(&account)
+        })?;
 
         // Now that we have the valid message storage PDA pubkey, we can get the full account data.
         let account = self
@@ -748,18 +755,24 @@ impl SealevelMailboxIndexer {
         nonce: u32,
     ) -> ChainResult<(Indexed<H256>, LogMeta)> {
         let discriminator = hyperlane_sealevel_mailbox::accounts::PROCESSED_MESSAGE_DISCRIMINATOR;
+        let nonce_bytes = nonce.to_le_bytes();
         let offset = 1 + 8 + 8; // the offset to get the `message_id` field
         let length = 32;
-        let accounts = self
-            .search_accounts_by_discriminator(&discriminator, nonce, offset, length)
-            .await?;
+        let accounts = search_accounts_by_discriminator(
+            self.rpc(),
+            &self.program_id,
+            &discriminator,
+            &nonce_bytes,
+            offset,
+            length,
+        )
+        .await?;
 
         debug!(account_len = ?accounts.len(), "Found accounts with processed message discriminator");
 
-        let valid_message_storage_pda_pubkey = self
-            .search_and_validate_account(accounts, |account| {
-                self.delivered_message_account(&account)
-            })?;
+        let valid_message_storage_pda_pubkey = search_and_validate_account(accounts, |account| {
+            self.delivered_message_account(&account)
+        })?;
 
         // Now that we have the valid delivered message storage PDA pubkey,
         // we can get the full account data.
@@ -797,67 +810,6 @@ impl SealevelMailboxIndexer {
             ChainCommunicationError::from_other_str("Could not find program address for message id")
         })?;
         Ok(expected_pubkey)
-    }
-
-    fn search_and_validate_account<F>(
-        &self,
-        accounts: Vec<(Pubkey, Account)>,
-        message_account: F,
-    ) -> ChainResult<Pubkey>
-    where
-        F: Fn(&Account) -> ChainResult<Pubkey>,
-    {
-        let mut valid_storage_pda_pubkey = Option::<Pubkey>::None;
-
-        for (pubkey, account) in accounts {
-            let expected_pubkey = message_account(&account)?;
-            if expected_pubkey == pubkey {
-                valid_storage_pda_pubkey = Some(pubkey);
-                break;
-            }
-        }
-
-        let valid_storage_pda_pubkey = valid_storage_pda_pubkey.ok_or_else(|| {
-            ChainCommunicationError::from_other_str("Could not find valid storage PDA pubkey")
-        })?;
-        Ok(valid_storage_pda_pubkey)
-    }
-
-    async fn search_accounts_by_discriminator(
-        &self,
-        discriminator: &[u8; 8],
-        nonce: u32,
-        offset: usize,
-        length: usize,
-    ) -> ChainResult<Vec<(Pubkey, Account)>> {
-        let target_message_account_bytes = &[&discriminator[..], &nonce.to_le_bytes()[..]].concat();
-        let target_message_account_bytes = base64::encode(target_message_account_bytes);
-
-        // First, find all accounts with the matching account data.
-        // To keep responses small in case there is ever more than 1
-        // match, we don't request the full account data, and just request
-        // the `unique_message_pubkey` field.
-        let memcmp = RpcFilterType::Memcmp(Memcmp {
-            // Ignore the first byte, which is the `initialized` bool flag.
-            offset: 1,
-            bytes: MemcmpEncodedBytes::Base64(target_message_account_bytes),
-            encoding: None,
-        });
-        let config = RpcProgramAccountsConfig {
-            filters: Some(vec![memcmp]),
-            account_config: RpcAccountInfoConfig {
-                encoding: Some(UiAccountEncoding::Base64),
-                data_slice: Some(UiDataSliceConfig { offset, length }),
-                commitment: Some(CommitmentConfig::finalized()),
-                min_context_slot: None,
-            },
-            with_context: Some(false),
-        };
-        let accounts = self
-            .rpc()
-            .get_program_accounts_with_config(&self.mailbox.program_id, config)
-            .await?;
-        Ok(accounts)
     }
 }
